restructuring folders, adding the edit task activity and add new task activity for admins
rescructured the Admin activity folder to have a folder with the main activities for the admin. as it began being quite a few activities. also added the edit task activity for admins with just the display of the user's task list for now. also added a add new task activity for the admin to add and create new task.