complete buttons, adding task to daily tasks
fixed the complete button on both user and admin side. when the admin clicks complete, the reward will be added into the bank in the database. if the task is not complete, the admin can click refresh. now the admin can also add tasks from the task list to the user's daily tasks list