adding background, refactoring, preperation
added the paper with lines to the backgrounds on the user side. refactored out all the java files that belongs to admin plan into an admin plan activity folder and all files belonging to admin task into a admin task activity. same with every file belonging to the bank activity. also made preperations for working on the bank activities